refactor: carousel컴포넌트 소문자로 변경
